Refactored 'stash_utils', and moved 'stash_utils/atm_stashlist.py' to 'amami/_atm_stashlist.py'
